Removed the global include file for the tests
The global include file is now removed and all is ready to make precompiled headers optional

+++ src/DiseaseSpreadSimulator/Person/Person.cpp
@@ -359,7 +359,7 @@ void DiseaseSpreadSimulation::Person::CheckNextMove(uint32_t currentTime, bool& 
 		if (isNewDay)
 		{
 			travelDays++;
-			if (Random::Percent<float>() <= (baseTravelReturnChance * (travelDays / static_cast<float>(3))))
+			if (Random::Percent<float>() <= (baseTravelReturnChance * (static_cast<float>(travelDays) / static_cast<float>(3))))
 			{
 				isTraveling = false;
 				travelDays = 0u;
